+++ src/Whim.Yaml.Tests/YamlLoader_LoadFocusIndicatorTests.cs
@@ -1,0 +1,6 @@
+namespace Whim.Yaml.Tests;
+
+public class YamlLoader_LoadFocusIndicatorTests
+{
+	// TODO
+}
