+++ src/Whim.Yaml/Whim.Yaml.csproj
@@ -31,6 +31,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\Whim.CommandPalette\Whim.CommandPalette.csproj" />
+		<ProjectReference Include="..\Whim.FocusIndicator\Whim.FocusIndicator.csproj" />
 		<ProjectReference Include="..\Whim.Gaps\Whim.Gaps.csproj" />
 		<ProjectReference Include="..\Whim\Whim.csproj" />
 	</ItemGroup>
