+++ src/Whim.Yaml/YamlLoader.cs
@@ -1,12 +1,8 @@
-using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using Corvus.Json;
-using Microsoft.UI;
-using Microsoft.UI.Xaml.Media;
 using Yaml2JsonNode;
 using YamlDotNet.RepresentationModel;
-using static Whim.Yaml.Schema;
 
 namespace Whim.Yaml;
 
@@ -17,7 +13,6 @@ public static class YamlLoader
 {
 	private const string JsonConfigFileName = "whim.config.json";
 	private const string YamlConfigFileName = "whim.config.yaml";
-	private const string HexColorStart = "#";
 
 	/// <summary>
 	/// Loads and applies the declarative configuration from a JSON or YAML file.
@@ -185,68 +180,5 @@ public static class YamlLoader
 					break;
 			}
 		}
-	}
-
-	internal static string SnakeToPascal(this string snake)
-	{
-		string[] parts = snake.Split('_');
-		StringBuilder builder = new(snake.Length);
-
-		foreach (string part in parts)
-		{
-			builder.Append(char.ToUpper(part[0]));
-			builder.Append(part.AsSpan(1));
-		}
-
-		return builder.ToString();
-	}
-
-	internal static BackdropType ParseBackdropType(this string backdropType) =>
-		backdropType switch
-		{
-			"none" => BackdropType.None,
-			"acrylic" => BackdropType.Acrylic,
-			"acrylic_thin" => BackdropType.AcrylicThin,
-			"mica" => BackdropType.Mica,
-			"mica_alt" => BackdropType.MicaAlt,
-			_ => BackdropType.None,
-		};
-
-	internal static WindowBackdropConfig ParseWindowBackdropConfig(WindowBackdropConfigEntity entity)
-	{
-		BackdropType backdropType = BackdropType.Mica;
-		bool alwaysShowBackdrop = true;
-
-		if (entity.BackdropType.AsOptional() is { } backdropTypeStr)
-		{
-			backdropType = ((string)backdropTypeStr).ParseBackdropType();
-		}
-
-		if (entity.AlwaysShowBackdrop.AsOptional() is { } alwaysShowBackdropValue)
-		{
-			alwaysShowBackdrop = alwaysShowBackdropValue;
-		}
-
-		return new WindowBackdropConfig(backdropType, alwaysShowBackdrop);
-	}
-
-
-
-	internal static Brush ParseBrush(this string brush)
-	{
-
-		if (brush.StartsWith(HexColorStart))
-		{
-			byte r = Convert.ToByte(brush.Substring(1, 2), 16);
-			byte g = Convert.ToByte(brush.Substring(3, 2), 16);
-			byte b = Convert.ToByte(brush.Substring(5, 2), 16);
-			byte a = Convert.ToByte(brush.Substring(7, 2), 16);
-			return new SolidColorBrush(ColorHelper.FromArgb(a, r, g, b));
-		}
-
-		string colorStr = brush.SnakeToPascal();
-		System.Drawing.Color color = System.Drawing.Color.FromName(colorStr);
-
-		return new SolidColorBrush(new Windows.UI.Color() { R = color.R, G = color.G, B = color.B });
 	}
 }
